+++ repositories/auth_repository.go
@@ -5,27 +5,18 @@ import (
 	"gorm.io/gorm"
 )
 
-type UserRepository interface {
-	CreateUser(user *models.User) error
-	GetUserByEmail(email string) (*models.User, error)
-}
-
-// userRepository 구조체
 type userRepository struct {
 	db *gorm.DB
 }
 
-// NewUserRepository - UserRepository 생성자 함수
 func NewUserRepository(db *gorm.DB) UserRepository {
 	return &userRepository{db: db}
 }
 
-// CreateUser - 새로운 사용자 생성
 func (r *userRepository) CreateUser(user *models.User) error {
 	return r.db.Create(user).Error
 }
 
-// GetUserByEmail - 이메일을 이용한 사용자 조회
 func (r *userRepository) GetUserByEmail(email string) (*models.User, error) {
 	var user models.User
 	result := r.db.Where("email = ?", email).First(&user)
